Separate email client from email request
