+++ app/services/notification/NotificationJob.scala
@@ -1,8 +1,8 @@
 package services.notification
 
-import com.amazonaws.services.simpleemail.AmazonSimpleEmailService
 import config.AMIableConfig
 import org.quartz.{Job, JobExecutionContext}
+import play.api.Configuration
 
 import scala.concurrent.ExecutionContext
 
@@ -12,8 +12,9 @@ class NotificationJob extends Job {
   override def execute(context: JobExecutionContext): Unit = {
     val schedulerContext = context.getScheduler.getContext
     val mailClient = schedulerContext.get("MailClient").asInstanceOf[AWSMailClient]
-    implicit val config = schedulerContext.get("AMIableConfig").asInstanceOf[AMIableConfig]
+    implicit val amiableConfig = schedulerContext.get("AMIableConfig").asInstanceOf[AMIableConfig]
     implicit val ec = schedulerContext.get("ExecutionContext").asInstanceOf[ExecutionContext]
+    implicit val configuration = schedulerContext.get("Configuration").asInstanceOf[Configuration]
     ScheduledNotificationRunner.run(mailClient)
   }
 }
